pyproject.toml
[project]
name = "dtom"
version = "0.1.0"
description = "Double Theory of Mind — Empirical grounding using TalkMoves dataset"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "openpyxl>=3.1",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "numpy>=1.24",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "scikit-learn>=1.3",
]

[tool.uv]
dev-dependencies = []

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dtom"]

[project.scripts]
dtom-pipeline = "dtom.analysis_pipeline:main"
dtom-llm = "dtom.llm_classifier:main"
dtom-ncte = "dtom.ncte_replication:main"
